Move classes to new manual package; make code compatible with that
