+++ spring-boot/src/test/java/org/dcsa/conformance/manual/ManualTestBase.java
@@ -1,22 +1,24 @@
-package org.dcsa.conformance.springboot;
-
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ObjectNode;
-import org.dcsa.conformance.core.party.HttpHeaderConfiguration;
-import org.slf4j.Logger;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-
-import java.util.List;
+package org.dcsa.conformance.manual;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.node.ObjectNode;
+import java.util.List;
+import org.dcsa.conformance.core.party.HttpHeaderConfiguration;
+import org.dcsa.conformance.sandbox.ConformanceWebuiHandler;
+import org.dcsa.conformance.springboot.ConformanceApplication;
+import org.junit.jupiter.api.BeforeEach;
+import org.slf4j.Logger;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+
 /** Base class which contains all API call methods and wiring needed to perform a manual test */
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT, classes = ConformanceApplication.class)
 public abstract class ManualTestBase {
   private static final String USER_ID = "unit-test";
 
@@ -29,16 +31,22 @@ public abstract class ManualTestBase {
   }
 
   @Autowired protected ConformanceApplication app;
+  private ConformanceWebuiHandler webuiHandler;
+
+  @BeforeEach
+  void setUp() {
+    webuiHandler = app.getWebuiHandler();
+  }
 
   void getAvailableStandards() {
     ObjectNode node = mapper.createObjectNode().put("operation", "getAvailableStandards");
-    JsonNode jsonNode = app.webuiHandler.handleRequest(USER_ID, node);
+    JsonNode jsonNode = webuiHandler.handleRequest(USER_ID, node);
     assertTrue(jsonNode.size() >= 6);
   }
 
   void getAllSandboxes() {
     ObjectNode node = mapper.createObjectNode().put("operation", "getAllSandboxes");
-    JsonNode jsonNode = app.webuiHandler.handleRequest(USER_ID, node);
+    JsonNode jsonNode = webuiHandler.handleRequest(USER_ID, node);
     assertTrue(jsonNode.size() >= 6);
   }
 
@@ -52,7 +60,7 @@ public abstract class ManualTestBase {
             .put("scenarioId", scenarioId)
             .put("actionId", actionId)
             .set("actionInput", textInputNode);
-    JsonNode jsonNode = app.webuiHandler.handleRequest(USER_ID, node);
+    JsonNode jsonNode = webuiHandler.handleRequest(USER_ID, node);
     assertTrue(jsonNode.isEmpty(), "Should be empty, found: " + jsonNode);
     waitForCleanSandboxStatus(sandbox);
   }
@@ -64,7 +72,7 @@ public abstract class ManualTestBase {
             .put("operation", "startOrStopScenario")
             .put("sandboxId", sandbox.sandboxId)
             .put("scenarioId", scenarioId);
-    assertTrue(app.webuiHandler.handleRequest(USER_ID, node).isEmpty());
+    assertTrue(webuiHandler.handleRequest(USER_ID, node).isEmpty());
   }
 
   void notifyAction(SandboxConfig sandbox) {
@@ -73,7 +81,7 @@ public abstract class ManualTestBase {
             .createObjectNode()
             .put("operation", "notifyParty")
             .put("sandboxId", sandbox.sandboxId);
-    assertTrue(app.webuiHandler.handleRequest(USER_ID, node).isEmpty());
+    assertTrue(webuiHandler.handleRequest(USER_ID, node).isEmpty());
     waitForCleanSandboxStatus(sandbox);
     waitForAsyncCalls(250L);
   }
@@ -84,7 +92,7 @@ public abstract class ManualTestBase {
             .createObjectNode()
             .put("operation", "completeCurrentAction")
             .put("sandboxId", sandbox.sandboxId);
-    JsonNode jsonNode = app.webuiHandler.handleRequest(USER_ID, node);
+    JsonNode jsonNode = webuiHandler.handleRequest(USER_ID, node);
     assertTrue(jsonNode.isEmpty(), "Should be empty, found: " + jsonNode);
     waitForCleanSandboxStatus(sandbox);
   }
@@ -125,7 +133,7 @@ public abstract class ManualTestBase {
             .put("operation", "getScenarioStatus")
             .put("sandboxId", sandbox.sandboxId)
             .put("scenarioId", scenarioId);
-    return app.webuiHandler.handleRequest(USER_ID, node);
+    return webuiHandler.handleRequest(USER_ID, node);
   }
 
   void waitForAsyncCalls(long delay) {
@@ -145,7 +153,7 @@ public abstract class ManualTestBase {
               .createObjectNode()
               .put("operation", "getSandboxStatus")
               .put("sandboxId", sandbox.sandboxId);
-      sandboxStatus = app.webuiHandler.handleRequest(USER_ID, node).toString();
+      sandboxStatus = webuiHandler.handleRequest(USER_ID, node).toString();
       if (sandboxStatus.contains("[]")) return;
       try {
         counter++;
@@ -175,7 +183,7 @@ public abstract class ManualTestBase {
             .put("externalPartyUrl", sandbox2.sandboxUrl)
             .put("externalPartyAuthHeaderName", sandbox2.sandboxAuthHeaderName)
             .put("externalPartyAuthHeaderValue", sandbox2.sandboxAuthHeaderValue);
-    assertTrue(app.webuiHandler.handleRequest(USER_ID, node).isEmpty());
+    assertTrue(webuiHandler.handleRequest(USER_ID, node).isEmpty());
 
     // Validate Sandbox 1 details
     node =
@@ -184,7 +192,7 @@ public abstract class ManualTestBase {
             .put("operation", "getSandbox")
             .put("sandboxId", sandbox1.sandboxId)
             .put("includeOperatorLog", true);
-    JsonNode jsonNode = app.webuiHandler.handleRequest(USER_ID, node);
+    JsonNode jsonNode = webuiHandler.handleRequest(USER_ID, node);
     assertTrue(jsonNode.has("id"));
     assertTrue(jsonNode.has("name"));
     assertTrue(jsonNode.has("operatorLog"));
@@ -202,14 +210,14 @@ public abstract class ManualTestBase {
             .put("testedPartyRole", sandbox.testedPartyRole)
             .put("isDefaultType", sandbox.isDefaultType)
             .put("sandboxName", sandbox.sandboxName);
-    JsonNode jsonNode = app.webuiHandler.handleRequest(USER_ID, node);
+    JsonNode jsonNode = webuiHandler.handleRequest(USER_ID, node);
     assertTrue(jsonNode.has("sandboxId"));
     String sandboxId = jsonNode.get("sandboxId").asText();
 
     // Get the sandbox config
     node =
         mapper.createObjectNode().put("operation", "getSandboxConfig").put("sandboxId", sandboxId);
-    jsonNode = app.webuiHandler.handleRequest(USER_ID, node);
+    jsonNode = webuiHandler.handleRequest(USER_ID, node);
     return mapper.convertValue(jsonNode, SandboxConfig.class);
   }
 
@@ -219,7 +227,7 @@ public abstract class ManualTestBase {
             .createObjectNode()
             .put("operation", "getScenarioDigests")
             .put("sandboxId", sandboxId);
-    JsonNode jsonNode = app.webuiHandler.handleRequest(USER_ID, node);
+    JsonNode jsonNode = webuiHandler.handleRequest(USER_ID, node);
     assertTrue(jsonNode.isArray());
     return mapper.convertValue(jsonNode, new TypeReference<>() {});
   }
